Add blogs app with models, views and tests
This commit introduces a new 'blogs' application with the initial set of models (Category and Blog) and corresponding tests. Additionally, admin interface for this application has been set up with customized admin sections for the new models. Django migrations files related to the introduced models are also included. Views have been created but are yet to be populated with content. Updates in blog_main settings.py and urls.py allow the handling of multimedia files.